refactor : 예약자 리스트 페이지 리팩토링
컬럼 내부로 수정
